refactor(webhook): update tests and snapshot for postback handlers
